Bumped versions -- SoDaUtils to 5.1.3 Bumpt SoDaLibs version to 5.5.1
This is an update to OptionsTest. It was calculating the error related
to a floating point input by (delta = a - b; error = delta / (a + b))
This doesn't work when we're testing for zero.

The new solution is similar in capability, but loses the ability to
produce nan when the two values match.